DiffableResultsController: Support updates
Currently, `UITableViewDiffableDataSource.apply()` does not refresh cells if the
cell was only updated. In here, we use the objects changed notifications to
forcefully update the snapshot using `reloadItems()` so that the
`UITableViewDiffableDataSource.apply()` will see that and reload the cell.

There is a radar for this issue: https://openradar.appspot.com/FB7588509.